Sud file decompression up and running
The sud audio input stream is working. This can be used to slot sud file decompression into anything that uses a Java AudioInputStream.

+++ X3/src/org/pamguard/x3/sud/WavFileHandler.java
@@ -28,7 +28,7 @@ import org.w3c.dom.NodeList;
  *
  */
 public class WavFileHandler implements ISudarDataHandler {
-	
+
 	/***Info on the wav file****/
 	private Integer fs;
 
@@ -42,7 +42,7 @@ public class WavFileHandler implements ISudarDataHandler {
 	 * The number of channels.
 	 */
 	private Integer nChan = 1;
-	
+
 	/***The chunk IDs associated with wav file writing***/
 
 	private int[] chunkIds;
@@ -51,7 +51,7 @@ public class WavFileHandler implements ISudarDataHandler {
 	 * The current sud file. 
 	 */
 	private File sudFile;
-	
+
 	/***Wav file writing using Java's in built Audio API****/
 
 	private String fileName;
@@ -62,16 +62,16 @@ public class WavFileHandler implements ISudarDataHandler {
 
 
 	private AudioInputStream audioInputStream;
-	
+
 	private Thread writeThread;
 
 	/***Xero drop out info****/
-	
+
 	long totalBytes = 0;
 
 	// the previous chunk to be written to the wav file. 
 	private Chunk lastChunk;
-	
+
 	//the value of the last error. 
 	private int lastError;
 
@@ -85,7 +85,7 @@ public class WavFileHandler implements ISudarDataHandler {
 
 	private Chunk firstChunk;
 
-	private boolean zeroFill;
+	private boolean zeroFill = true; 
 
 	private LogFileStream logFile;
 
@@ -96,28 +96,35 @@ public class WavFileHandler implements ISudarDataHandler {
 
 	private Integer nBits; 
 
-	
+	/**
+	 * True to save the wav files. 
+	 */
+	private boolean saveWav = true;
+
+
 	//the difference between the sample count and the device's on board clock before a correction is made
 	public static double timeErrorWarningThreshold = 0.04; //%
 
 
+	public WavFileHandler(SudParams filePath, String ftype) {
+		this.sudFile = new File(filePath.getSudFilePath());
 
-	public WavFileHandler(String filePath, String ftype) {
-		this.sudFile = new File(filePath);
+		this.fileName = filePath.getOutFilePath();
+		this.saveWav = filePath.saveWav;
+		this.zeroFill = filePath.zeroPad;
 
-		this.fileName = FilenameUtils.removeExtension(sudFile.getName());
 		this.ftype=ftype; 
 	}
-	
+
 	int count=0;
 
 	@Override
 	public void processChunk(Chunk sudChunk) {
-		
+
 		//System.out.println("Process wav file: " + sudChunk.buffer.length + "  " + sudChunk.buffer[0]);
-		
+
 		//create the audio file. 
-		if (audioFile==null) {
+		if (audioFile==null && saveWav) {
 			audioFile = new File(fileName + fileSuffix + ".wav");
 			if (audioFile.exists()) {
 				audioFile.delete();
@@ -125,29 +132,29 @@ public class WavFileHandler implements ISudarDataHandler {
 			writeThread = new Thread(new WriteThread());
 			writeThread.start();
 		}
-		
+
 		try
 		{
 			//Test to see whether we need to zero fill drop outs. 
 			if (lastChunk==null) {
 				//the first chunk - nothing we can do here. 
 				this.firstChunk = sudChunk; 
-				
+
 			}
 			else {
 				int elapsedTimeS = (sudChunk.chunkHeader.TimeS - lastChunk.chunkHeader.TimeS);
-				
+
 				//the time between chunks in micoseconds
 				long elapsedTimeUs = (long)((elapsedTimeS * 1000000) + (sudChunk.chunkHeader.TimeOffsetUs - lastChunk.chunkHeader.TimeOffsetUs));
-				
+
 				//the time between chunks calculated using samples. 
 				long calculatedTime = (long)((long)lastChunk.chunkHeader.SampleCount * 1000000 / fs);
-				
+
 				//these two times should be the same, 
 				int error = (int) (elapsedTimeUs - calculatedTime); //this is the total number of samples to add
 				cumulativeTimeErrorUs += error; 
 				lastChunk = sudChunk;
-				
+
 				if ((timeCheck > 0) && (Math.abs(error) > (calculatedTime * timeErrorWarningThreshold))) {
 					//log warning
 					double t = (double)(sudChunk.chunkHeader.TimeS - firstChunk.chunkHeader.TimeS) + 
@@ -157,8 +164,9 @@ public class WavFileHandler implements ISudarDataHandler {
 							//String.format("Sampling Gap {0} us at sample {1} ({2} s), chunk {3}", error, cumulativeSamples, t, chunkCount);
 							logFile.writeXML(this.chunkIds[0], "WavFileHandler", "Info", String.format("Sampling Gap {0} us at sample {1} ({2} s), chunk {3}", error, cumulativeSamples, t, chunkCount));
 							if (zeroFill) {
-								
-								int samplesToAdd = (int)(error * fs / 1000000);
+
+								//System.out.println("Error: " + error + " " + nChan ); 
+								int samplesToAdd = (int)(error * (fs / 1000000));
 								byte[] fill = new byte[samplesToAdd * 2 * nChan];
 								pipedOutputStream.write(fill);
 
@@ -178,10 +186,12 @@ public class WavFileHandler implements ISudarDataHandler {
 				}
 				lastError = error; 
 			}
-			
+
 			//System.out.println("Write wav: " + ++count + " " + sudChunk.chunkHeader.HeaderCrc );
-						
-			pipedOutputStream.write(sudChunk.buffer);
+
+			if (saveWav) {
+				pipedOutputStream.write(sudChunk.buffer);
+			}
 			lastChunk = sudChunk;
 
 
@@ -190,12 +200,12 @@ public class WavFileHandler implements ISudarDataHandler {
 		{
 			e.printStackTrace();
 		}
-		
+
 	}
 
 	@Override
 	public void close() {
-		
+
 		if (this.lastChunk!=null) {
 
 			// the format of your date
@@ -203,7 +213,7 @@ public class WavFileHandler implements ISudarDataHandler {
 			// give a timezone reference for formatting (see comment at the bottom)
 			SimpleDateFormat sdfLocal = new java.text.SimpleDateFormat("yyyy-MM-dd HH:mm:ss z"); 
 			sdf.setTimeZone(TimeZone.getDefault()); 
-			
+
 			long recordPeriod = lastChunk.chunkHeader.TimeS - firstChunk.chunkHeader.TimeS;
 			recordPeriod *= 1000000;
 			recordPeriod += (long)lastChunk.chunkHeader.TimeOffsetUs - firstChunk.chunkHeader.TimeOffsetUs;
@@ -217,37 +227,39 @@ public class WavFileHandler implements ISudarDataHandler {
 			logFile.writeXML(chunkIds[0], "WavFileHandler", "CumulativeSamplingGap", String.format("%d us", cumulativeTimeErrorUs));
 			logFile.writeXML(chunkIds[0], "WavFileHandler", "SampleCount", String.format("%d", cumulativeSamples));
 		}
-		
-		//System.out.println("Close the file:"); 
-		try {
-			pipedOutputStream.flush();
-			pipedOutputStream.close();
-			// don't close audioInputStream - the actual bit that writes. It will 
-			// close itself when it's done in it's own thread. But we DO need to wait for 
-			// that to happen. 
-//			audioInputStream.
-//			audioInputStream.close();
-		} catch (IOException e) {
-			e.printStackTrace();
-		}
-		/*
-		 *  Wait for the write thread to complete (but don't wait forever !
-		 * 
-		 */
-		int nKips = 0;
-		for (int i = 0; i < 50; i++) {
-			if (audioInputStream == null) {
-				break;
-			}
+
+		if (saveWav) {
+			//System.out.println("Close the file:"); 
 			try {
-				Thread.sleep(100);
-				nKips++;
-			} catch (InterruptedException e) {
+				pipedOutputStream.flush();
+				pipedOutputStream.close();
+				// don't close audioInputStream - the actual bit that writes. It will 
+				// close itself when it's done in it's own thread. But we DO need to wait for 
+				// that to happen. 
+				//			audioInputStream.
+				//			audioInputStream.close();
+			} catch (IOException e) {
 				e.printStackTrace();
 			}
-		}
-		if (nKips >= 49) {
-			System.out.println(String.format("Warning !!! Waited %d milliseconds for file write to complete.", nKips*100));
+			/*
+			 *  Wait for the write thread to complete (but don't wait forever !
+			 * 
+			 */
+			int nKips = 0;
+			for (int i = 0; i < 50; i++) {
+				if (audioInputStream == null) {
+					break;
+				}
+				try {
+					Thread.sleep(100);
+					nKips++;
+				} catch (InterruptedException e) {
+					e.printStackTrace();
+				}
+			}
+			if (nKips >= 49) {
+				System.out.println(String.format("Warning !!! Waited %d milliseconds for file write to complete.", nKips*100));
+			}
 		}
 	}
 
@@ -255,15 +267,15 @@ public class WavFileHandler implements ISudarDataHandler {
 	public void init(LogFileStream inputStream, String innerXml, int id) {
 		this.logFile = inputStream;
 		this.chunkIds = new int[]{id};
-		
+
 		Document doc = XMLFileHandler.convertStringToXMLDocument(innerXml.trim());
 
 		NodeList nodeList = doc.getElementsByTagName("CFG");
-		
+
 		HashMap<String, String> nodeContent = XMLUtils.getInnerNodeContent(new String[] {"FS", "SUFFIX", "TIMECHK", "CHANNEL", "NCHS", "NBITS"},  nodeList);
-//		for (int i=0; i<nodeContent.size(); i++) {
-//			System.out.println(nodeContent.values().toArray()[i]);
-//		}
+		//		for (int i=0; i<nodeContent.size(); i++) {
+		//			System.out.println(nodeContent.values().toArray()[i]);
+		//		}
 
 		channel = -1;
 		nChan = 1;
@@ -271,10 +283,10 @@ public class WavFileHandler implements ISudarDataHandler {
 		cumulativeSamples = 0;
 		chunkCount = 0;
 		lastChunk = null;
-		
+
 		fs = Integer.valueOf(nodeContent.get("FS"));
 		fileSuffix = nodeContent.get("SUFFIX");
-		
+
 		nBits = Integer.valueOf(nodeContent.get("NBITS"));
 
 		if (nodeContent.get("TIMECHK")!=null) timeCheck = Integer.valueOf(nodeContent.get("TIMECHK"));
@@ -283,18 +295,20 @@ public class WavFileHandler implements ISudarDataHandler {
 
 		//create the audio format. 
 		AudioFormat audioFormat = new AudioFormat(fs, 16,nChan, true, false);
-		
-		///create the wav writer
-		PipedInputStream pipedInputStream;
-		pipedOutputStream = null;
-		try {
-			pipedInputStream = new PipedInputStream();
-			pipedOutputStream = new PipedOutputStream(pipedInputStream);
-			audioInputStream = new AudioInputStream(pipedInputStream, audioFormat, AudioSystem.NOT_SPECIFIED);
 
-		}
-		catch (IOException Ex) {
-			Ex.printStackTrace();
+		if (saveWav) {
+			///create the wav writer
+			PipedInputStream pipedInputStream;
+			pipedOutputStream = null;
+			try {
+				pipedInputStream = new PipedInputStream();
+				pipedOutputStream = new PipedOutputStream(pipedInputStream);
+				audioInputStream = new AudioInputStream(pipedInputStream, audioFormat, AudioSystem.NOT_SPECIFIED);
+
+			}
+			catch (IOException Ex) {
+				Ex.printStackTrace();
+			}
 		}
 	}
 
@@ -302,7 +316,7 @@ public class WavFileHandler implements ISudarDataHandler {
 	public int[] getChunkID() {
 		return chunkIds;
 	}
-	
+
 	class WriteThread implements Runnable {
 
 		@Override
@@ -310,7 +324,7 @@ public class WavFileHandler implements ISudarDataHandler {
 			writeData();
 		}
 	}
-	
+
 	/**
 	 * Called within the write thread, this does not return
 	 * until the pipes get closed. 
@@ -365,6 +379,20 @@ public class WavFileHandler implements ISudarDataHandler {
 	public int getNChannels() {
 		return nChan;
 	}
-	
+
+	//	
+	//	public void setSudarParams(SudParams sudParams) {
+	//		this.zeroFill = sudParams.zeroPad;
+	//		this.saveWav = sudParams.saveWav; 
+	//
+	//		if (sudParams.saveFolder!=null) {
+	//			String sudName = FilenameUtils.removeExtension(sudFile.getName());
+	//			//the audioFile will be reset. 
+	//			this.audioFile =null; 
+	//			this.fileName = sudParams.saveFolder + File.pathSeparator + sudName;
+	//		}
+	//
+	//	}
+
 
 }
